fix: refactor stream.
