+++ protocol-units/da/movement/protocol/util/src/blob/ir/blob.rs
@@ -1,0 +1,129 @@
+use crate::blob::ir::data::InnerSignedBlobV1Data;
+use crate::blob::ir::id::Id;
+use ecdsa::{
+	elliptic_curve::{
+		generic_array::ArrayLength,
+		ops::Invert,
+		point::PointCompression,
+		sec1::{FromEncodedPoint, ModulusSize, ToEncodedPoint},
+		subtle::CtOption,
+		AffinePoint, CurveArithmetic, FieldBytesSize, PrimeCurve, Scalar,
+	},
+	hazmat::{DigestPrimitive, SignPrimitive, VerifyPrimitive},
+	signature::{digest::Digest, DigestVerifier},
+	SignatureSize, VerifyingKey,
+};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct InnerSignedBlobV1 {
+	pub data: InnerSignedBlobV1Data,
+	pub signature: Vec<u8>,
+	pub signer: Vec<u8>,
+	pub id: Id,
+}
+
+impl InnerSignedBlobV1 {
+	pub fn try_verify<C>(&self) -> Result<(), anyhow::Error>
+	where
+		C: PrimeCurve + CurveArithmetic + DigestPrimitive + PointCompression,
+		Scalar<C>: Invert<Output = CtOption<Scalar<C>>> + SignPrimitive<C>,
+		SignatureSize<C>: ArrayLength<u8>,
+		AffinePoint<C>: FromEncodedPoint<C> + ToEncodedPoint<C> + VerifyPrimitive<C>,
+		FieldBytesSize<C>: ModulusSize,
+	{
+		let mut hasher = C::Digest::new();
+		hasher.update(self.data.blob.as_slice());
+		hasher.update(&self.data.timestamp.to_be_bytes());
+		hasher.update(self.id.as_slice());
+
+		let verifying_key = VerifyingKey::<C>::from_sec1_bytes(self.signer.as_slice())?;
+		let signature = ecdsa::Signature::from_bytes(self.signature.as_slice().into())?;
+
+		match verifying_key.verify_digest(hasher, &signature) {
+			Ok(_) => Ok(()),
+			Err(_) => Err(anyhow::anyhow!("Failed to verify signature")),
+		}
+	}
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub enum IntermediateBlobRepresentation {
+	SignedV1(InnerSignedBlobV1),
+}
+
+impl From<InnerSignedBlobV1> for IntermediateBlobRepresentation {
+	fn from(inner: InnerSignedBlobV1) -> Self {
+		IntermediateBlobRepresentation::SignedV1(inner)
+	}
+}
+
+impl IntermediateBlobRepresentation {
+	pub fn blob(&self) -> &[u8] {
+		match self {
+			IntermediateBlobRepresentation::SignedV1(inner) => inner.data.blob.as_slice(),
+		}
+	}
+
+	pub fn signature(&self) -> &[u8] {
+		match self {
+			IntermediateBlobRepresentation::SignedV1(inner) => inner.signature.as_slice(),
+		}
+	}
+
+	pub fn timestamp(&self) -> u64 {
+		match self {
+			IntermediateBlobRepresentation::SignedV1(inner) => inner.data.timestamp,
+		}
+	}
+
+	pub fn signer(&self) -> &[u8] {
+		match self {
+			IntermediateBlobRepresentation::SignedV1(inner) => inner.signer.as_slice(),
+		}
+	}
+
+	pub fn signer_hex(&self) -> String {
+		hex::encode(self.signer())
+	}
+
+	pub fn id(&self) -> &[u8] {
+		match self {
+			IntermediateBlobRepresentation::SignedV1(inner) => inner.id.as_slice(),
+		}
+	}
+
+	pub fn verify_signature<C>(&self) -> Result<(), anyhow::Error>
+	where
+		C: PrimeCurve + CurveArithmetic + DigestPrimitive + PointCompression,
+		Scalar<C>: Invert<Output = CtOption<Scalar<C>>> + SignPrimitive<C>,
+		SignatureSize<C>: ArrayLength<u8>,
+		AffinePoint<C>: FromEncodedPoint<C> + ToEncodedPoint<C> + VerifyPrimitive<C>,
+		FieldBytesSize<C>: ModulusSize,
+	{
+		match self {
+			IntermediateBlobRepresentation::SignedV1(inner) => inner.try_verify::<C>(),
+		}
+	}
+}
+
+#[cfg(test)]
+pub mod test {
+
+	use super::*;
+	use ecdsa::SigningKey;
+
+	#[test]
+	fn test_cannot_change_id_and_verify() -> Result<(), anyhow::Error> {
+		let blob = InnerSignedBlobV1Data::new(vec![1, 2, 3], 123);
+		let signing_key = SigningKey::<k256::Secp256k1>::random(&mut rand::thread_rng());
+		let signed_blob = blob.try_to_sign(&signing_key)?;
+
+		let mut changed_blob = signed_blob.clone();
+		changed_blob.id = Id::new(vec![1, 2, 3, 4]);
+
+		assert!(changed_blob.try_verify::<k256::Secp256k1>().is_err());
+
+		Ok(())
+	}
+}
