+++ protocol-units/da/movement/protocol/util/src/blob/mod.rs
@@ -1,0 +1,1 @@
+pub mod ir;
