+++ protocol-units/da/movement/protocol/util/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod config;
 pub use config::*;
-pub mod ir_blob;
+pub mod blob;
