+++ protocol-units/da/movement/protocol/verifier/src/permissioned_signers/mod.rs
@@ -17,7 +17,7 @@ use ecdsa::{
 	hazmat::{DigestPrimitive, SignPrimitive, VerifyPrimitive},
 	SignatureSize,
 };
-use movement_da_util::ir_blob::IntermediateBlobRepresentation;
+use movement_da_util::blob::ir::blob::IntermediateBlobRepresentation;
 use std::sync::Arc;
 
 /// A verifier of Celestia blobs for permissioned signers
@@ -69,7 +69,11 @@ where
 	AffinePoint<C>: FromEncodedPoint<C> + ToEncodedPoint<C> + VerifyPrimitive<C>,
 	FieldBytesSize<C>: ModulusSize,
 {
-	async fn verify(&self, blob: CelestiaBlob, height: u64) -> Result<Verified<IntermediateBlobRepresentation>, Error> {
+	async fn verify(
+		&self,
+		blob: CelestiaBlob,
+		height: u64,
+	) -> Result<Verified<IntermediateBlobRepresentation>, Error> {
 		let verified_blob = self.celestia.verify(blob, height).await?;
 		self.known_signers.verify(verified_blob.into_inner(), height).await
 	}
