+++ protocol-units/da/movement/providers/celestia/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod blob;
 pub mod da;
+pub mod verifier;
